Prepare for removal of workarould in Firefox's LookAndFeel code
In widget/nsXPLookAndFeel.cpp, an "#ifdef MOZ_WIDGET_GTK" currently
returns "SystemColorScheme()" instead of defaulting to
ColorScheme::Light. This logic however might be refactored away.

This adds a "color-scheme" meta tag to the sidebar page, so that the
page of the sidebar can fall through the
"PreferredColorScheme/suppportsDark/supportsLight" logic
instead of defaulting to a light theme in the future.